Hide equipment panel in overview if its empty, move equipment overlay to the tab instead of the underlying component
